pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "mcpgoogle"
version = "0.2.0"
description = "FastMCP server for Google services (Gmail, Calendar) with OMA Backend integration"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.13.0.2",
    "google-api-python-client>=2.149.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.0",
    "python-dotenv>=1.0.1",
    "httpx>=0.27.0",  # For OMA backend HTTP client
]

[project.optional-dependencies]
dev = [
    "ruff>=0.14.2",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]
